feat: 반복일정 구현 및 도커 환경세팅
