+++ .gitignore
@@ -1,4 +1,6 @@
 .gradle
+datasource.env
+mysql.env
 build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
