+++ module-application/src/main/java/com/studdit/schedule/controller/ScheduleController.java
@@ -10,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.web.bind.annotation.*;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -32,7 +31,8 @@ public class ScheduleController {
             @RequestParam(required = false, defaultValue = "month") String view,
             @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) LocalDateTime date
     ) {
-        return ApiResponse.ok(scheduleService.findSchedules(username, view, date)); // to-do
+//        return ApiResponse.ok(scheduleService.findSchedules(username, view, date)); // to-do
+        return ApiResponse.ok(null);
     }
 
     @PutMapping("/schedules/{id}")
