+++ module-application/src/main/java/com/studdit/schedule/service/ScheduleService.java
@@ -1,0 +1,137 @@
+package com.studdit.schedule.service;
+
+
+import com.studdit.schedule.domain.RecurrenceRule;
+import com.studdit.schedule.domain.Schedule;
+import com.studdit.schedule.domain.ScheduleInstance;
+import com.studdit.schedule.enums.ScheduleViewType;
+import com.studdit.schedule.repository.*;
+import com.studdit.schedule.request.RecurrenceRuleCreateServiceRequest;
+import com.studdit.schedule.request.ScheduleCreateServiceRequest;
+import com.studdit.schedule.request.ScheduleModifyServiceRequest;
+import com.studdit.schedule.response.ScheduleResponse;
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.Duration;
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ScheduleService {
+
+    private final ScheduleRepository scheduleRepository;
+    private final ScheduleInstanceRepsitory scheduleInstanceRepsitory;
+    private final RecurrenceRuleRepository recurrenceRuleRepository;
+    private final ScheduleInstanceGenerator instanceGenerator;
+
+    // Business
+    @Transactional
+    public ScheduleResponse createSchedule(ScheduleCreateServiceRequest request) {
+
+        // 1. 마스터 일정 생성
+        Schedule schedule = request.toEntity();
+        Schedule savedSchedule = scheduleRepository.save(schedule);
+
+        // 2. 일정 인스턴스 생성
+        createScheduleInstances(savedSchedule.getId(), request);
+        return ScheduleResponse.of(savedSchedule);
+    }
+
+
+    @Transactional
+    public ScheduleResponse modifySchedule(ScheduleModifyServiceRequest request) {
+
+        Schedule requestSchedule = request.toEntity();
+        Schedule schedule = scheduleRepository.findById(request.getId())
+                .orElseThrow(() -> new EntityNotFoundException("해당 ID의 일정을 찾을 수 없습니다."));
+
+        schedule.update(requestSchedule);
+
+        return ScheduleResponse.of(schedule);
+    }
+/*
+    public List<ScheduleResponse> findSchedules(String username, String view, LocalDateTime date) {
+        ScheduleViewType viewType = validateAndParseViewType(view);
+        DateRange dateRange = calculateDateRange(viewType, date);
+
+        List<Schedule> schedules = scheduleRepository.findByDateRange(dateRange.getStart(), dateRange.getEnd());
+
+        return schedules.stream()
+                .map(ScheduleResponse::of)
+                .collect(Collectors.toList());
+    }
+    /*
+ */
+
+    @Transactional
+    public ScheduleResponse deleteSchedule(Long id) {
+        Schedule schedule = scheduleRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("해당 ID의 일정을 찾을 수 없습니다."));
+
+        scheduleRepository.delete(schedule);
+
+        return null;
+    }
+
+    // implementation
+
+    private void createScheduleInstances(Long scheduleId, ScheduleCreateServiceRequest request) {
+        if (request.getRecurrenceRuleCreateServiceRequest() == null) {
+            createSingleSchedule(scheduleId, request);
+        } else {
+            createRecurringSchedule(scheduleId, request);
+        }
+    }
+
+    private void createSingleSchedule(Long scheduleId, ScheduleCreateServiceRequest request) {
+
+    }
+
+    private void createRecurringSchedule(Long scheduleId, ScheduleCreateServiceRequest request) {
+        RecurrenceRule recurrenceRule = createRecurrenceRule(scheduleId, request.getRecurrenceRuleCreateServiceRequest());
+        recurrenceRuleRepository.save(recurrenceRule);
+        Duration duration = Duration.between(request.getStartDateTime(), request.getEndDateTime());
+        List<ScheduleInstance> instances = instanceGenerator.generateInstances(recurrenceRule, request.getStartDateTime(), duration);
+        scheduleInstanceRepsitory.saveAll(instances);
+    }
+
+    // to-do RecurrenceRule의 객체를 만드는 책임
+    private RecurrenceRule createRecurrenceRule(Long scheduleId, RecurrenceRuleCreateServiceRequest recurrenceRule) {
+        return RecurrenceRule.builder()
+                .scheduleId(scheduleId)
+                .frequency(recurrenceRule.getFrequency())
+                .type(recurrenceRule.getType())
+                .endDate(recurrenceRule.getEndDate())
+                .byWeekday(recurrenceRule.getByWeekday())
+                .byMonthday(recurrenceRule.getByMonthday())
+                .byMonth(recurrenceRule.getByMonth())
+                .maxOccurrences(recurrenceRule.getMaxOccurrences())
+                .build();
+    }
+
+    private ScheduleViewType validateAndParseViewType(String view) {
+        try {
+            return ScheduleViewType.valueOf(view.toUpperCase());
+        } catch (IllegalArgumentException e) {
+            throw new IllegalArgumentException("지원하지 않는 뷰 타입입니다: " + view);
+        }
+    }
+
+    private DateRange calculateDateRange(ScheduleViewType viewType, LocalDateTime date) {
+        switch (viewType) {
+            case DAY:
+                return DateRange.ofDay(date);
+            case WEEK:
+                return DateRange.ofWeek(date);
+            case MONTH:
+                return DateRange.ofMonth(date);
+            default:
+                throw new IllegalArgumentException("지원하지 않는 뷰 타입입니다: " + viewType);
+        }
+    }
+}
