+++ module-common/src/main/java/com/studdit/schedule/enums/RecurrenceType.java
@@ -1,0 +1,13 @@
+package com.studdit.schedule.enums;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum RecurrenceType {
+    DAILY("매일"),
+    WEEKLY("매주"),
+    MONTHLY("매월"),
+    YEARLY("매년");
+
+    private final String text;
+}
