+++ module-common/src/main/java/com/studdit/schedule/enums/ScheduleViewType.java
