+++ module-common/src/main/java/com/studdit/schedule/enums/VariantType.java
@@ -1,0 +1,12 @@
+package com.studdit.schedule.enums;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum VariantType {
+    ORIGINAL("원본"),
+    MODIFIED("수정됨"),
+    CANCELLED("취소됨");
+
+    private final String text;
+}
