+++ module-db/build.gradle
@@ -1,0 +1,25 @@
+apply plugin: 'io.spring.dependency-management'
+apply plugin: 'java-library'
+
+dependencyManagement {
+    imports {
+        mavenBom 'org.springframework.boot:spring-boot-dependencies:3.4.4'
+    }
+}
+
+dependencies {
+    api 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    implementation project(':module-common')
+
+    // 도메인 로직에 필요한 핵심 스프링 의존성
+    compileOnly 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
+    runtimeOnly 'com.h2database:h2'
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+
+    tasks.named('test') {
+        useJUnitPlatform()
+    }
+}
