+++ module-db/src/main/java/com/studdit/schedule/domain/RecurringVariant.java
@@ -1,0 +1,32 @@
+package com.studdit.schedule.domain;
+
+import com.studdit.BaseEntity;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+
+import static jakarta.persistence.GenerationType.*;
+
+@Getter
+@NoArgsConstructor
+@Entity
+public class RecurringVariant extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = IDENTITY)
+    private Long id;
+
+    private Long scheduleId;
+
+    private LocalDate originalDate;
+
+    private Long variantInstanceId;
+
+    private Boolean isCancelled = false;
+}
+
