+++ module-db/src/main/java/com/studdit/schedule/repository/ScheduleRepository.java
@@ -1,0 +1,15 @@
+package com.studdit.schedule.repository;
+
+import com.studdit.schedule.domain.Schedule;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Repository
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
+
+}
