+++ settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'studdit'
-include('module-application', 'module-common', 'module-domain', 'module-independant', 'module-internal')
+include('module-application', 'module-common', 'module-db', 'module-independant', 'module-internal')
